requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
cffi==1.15.1
charset-normalizer==2.1.0
citeproc-py==0.6.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cognitiveatlas==0.1.9
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.30
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-braces==1.15.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-datatables-view==1.20.0
django-file-resubmit==0.5.2
django-filter==22.1
django-guardian==2.4.0
django-oauth-toolkit==2.1.0
django-polymorphic==3.1.0
django-sendfile==0.3.11
django-taggit==3.0.0
djangorestframework==3.13.1
duecredit==0.9.1
fonttools==4.34.4
frozendict==2.3.2
future==0.18.2
fuzzywuzzy==0.18.0
h5py==3.7.0
html5lib==1.1
idna==3.3
imageio==2.19.3
indexed-gzip==1.6.13
isodate==0.6.1
joblib==1.1.0
jwcrypto==1.3.1
kiwisolver==1.4.3
kombu==5.2.4
lxml==4.9.1
matplotlib==3.4.3
mpmath==1.2.1
networkx==2.8.4
nibabel==4.0.1
nidmfsl==2.2.0
nidmresults==2.1.0
nilearn==0.9.1
NiMARE==0.0.11
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
patsy==0.5.2
Pillow==9.2.0
prompt-toolkit==3.0.30
prov==2.0.0
psycopg==3.0.15
psycopg2==2.9.3
pycortex==1.2.5
pycparser==2.21
PyJWT==2.4.0
PyLD==2.0.3
PyMARE==0.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
rdflib==6.1.1
rdflib-jsonld==0.6.2
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
statsmodels==0.13.2
sympy==1.10.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.0
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
wrapt==1.14.1
